[Gold IV] Title: 카드 정렬하기, Time: 376 ms, Memory: 25240 KB -BaekjoonHub

+++ 백준/Gold/1715. 카드 정렬하기/카드 정렬하기.java
@@ -5,17 +5,17 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         int N = Integer.parseInt(br.readLine());
-        PriorityQueue<Long> pq = new PriorityQueue<>();
-        long res = 0;
+        PriorityQueue<Integer> pq = new PriorityQueue<>();
+        int res = 0;
 
         for (int i = 0; i < N; i++) {
-            pq.add(Long.parseLong(br.readLine()));
+            pq.add(Integer.parseInt(br.readLine()));
         }
         br.close();
         
         while (pq.size() > 1) {
-            long a = pq.poll();
-            long b = pq.poll();
+            int a = pq.poll();
+            int b = pq.poll();
             res += (a + b);
             pq.add(a+b);
         }
